+++ build.gradle.kts
@@ -19,13 +19,11 @@ buildscript {
 }
 
 plugins {
-    id("com.google.dagger.hilt.android").version(libs.versions.hilt).apply(false)
     id("io.gitlab.arturbosch.detekt").version(libs.versions.detektGradlePlugin)
     id("org.jmailen.kotlinter").version(libs.versions.kotlinter).apply(false)
 }
 
 apply(from = "buildscripts/githooks.gradle")
-apply(from = "buildscripts/setup.gradle")
 apply(from = "buildscripts/versionsplugin.gradle")
 
 subprojects {
